+++ src/game/mod.rs
@@ -1,4 +1,3 @@
-pub mod common;
 pub mod cycle;
 pub mod enemy;
 pub mod player;
